+++ src/H2DE_loader.cpp
@@ -10,7 +10,3 @@ SDL_Texture* H2DE_Loader::loadTexture(SDL_Renderer* renderer, const char* file) 
 Mix_Chunk* H2DE_Loader::loadSound(const char* file) {
     return Mix_LoadWAV(file);
 }
-
-TTF_Font* H2DE_Loader::loadFont(const char* file) {
-    return TTF_OpenFont(file, 24);
-}